requirements.txt
# Dissect - Dependencies

# Telegram Bot
python-telegram-bot==21.0

# Groq AI (fast, generous free tier)
groq>=0.4.0

# Environment variables
python-dotenv==1.0.0

# Async HTTP (enrichment APIs)
aiohttp>=3.9.0

# SQLite async support
aiosqlite>=0.19.0
